feat: move BcryptAdapter to a folder
