+++ src/infra/cryptography/jwt-adapter/index.ts
@@ -1,0 +1,3 @@
+export class JwtAdapter {
+  //
+}
